requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
requests>=2.32.0
streamlit==1.36.0
pandas==2.2.2
PyYAML==6.0.1
openai==1.106.1
google-analytics-data==0.18.11
google-cloud-bigquery==3.25.0
google-cloud-pubsub==2.20.0
google-cloud-storage==2.17.0
google-cloud-secret-manager==2.20.2
ndjson==0.3.1
tqdm==4.66.4
plotly==5.22.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-exporter-otlp==1.25.0
loguru==0.7.2
